Stop using dotnet tool to install
